+++ server/src/bin.rs
@@ -16,6 +16,7 @@ mod jsonerrors;
 mod process;
 mod routes;
 pub mod serve;
+mod timer;
 // #[cfg(feature = "search")]
 mod search;
 #[cfg(test)]