+++ CHANGELOG.md
@@ -6,6 +6,7 @@ By far most changes relate to `atomic-server`, so if not specified, assume the c
 ## [unreleased]
 
 - Correct 404 status code
+- Server-Timings header #256
 
 ## [v0.30.0] - 2021-12-22